Цена кофе пересчитана, температура настроена, кофемашина интегрирована в торгомат, подключен GUI. Именно после его подключения автомат и сломался.

+++ src/Arabica.java
@@ -1,5 +1,5 @@
 public class Arabica extends Coffee{
-    public Arabica(String name, double price, int value) {
+    public Arabica(String name, double price, double value) {
         super(name, price, value);
     }
 }
