+++ src/Milk.java
@@ -1,7 +1,7 @@
 public class Milk extends Product{
     private double temperature;
 
-    public Milk(String name, double price, int value) {
+    public Milk(String name, double price, double value) {
         super(name, price, value);
         temperature = 3;
     }
